Feat: JWT 구현

+++ controller/CartController.js
@@ -14,7 +14,8 @@ const checkExistValues = async (connection, values) => {
 
 const addToCart = async (req, res) => {
     const connection = await conn.getConnection();
-    const { bookId, quantity, userId } = camelcaseKeys(req.body);
+    const { bookId, quantity } = camelcaseKeys(req.body);
+    const userId = req.userId;
 
     const sqlInsertCart = 'insert into cartItems (book_id, quantity, user_id) values (?, ?, ?)';
     const sqlSelectCart = 'select * from cartItems where user_id = ? and book_id = ?';
@@ -23,13 +24,7 @@ const addToCart = async (req, res) => {
     const existValues = [userId, bookId];
 
     try {
-        const { user_exists, book_exists } = await checkExistValues(connection, existValues);
-
-        if (!user_exists) {
-            return res.status(StatusCodes.NOT_FOUND).json({
-                message: '존재하지 않는 유저입니다.',
-            });
-        }
+        const { book_exists } = await checkExistValues(connection, existValues);
 
         if (!book_exists) {
             return res.status(StatusCodes.NOT_FOUND).json({
@@ -77,9 +72,9 @@ const addToCart = async (req, res) => {
 
 const getCartItems = async (req, res) => {
     const connection = await conn.getConnection();
-    const { userId, selected } = camelcaseKeys(req.body);
+    const userId = req.userId;
+    const { selected } = camelcaseKeys(req.body);
 
-    const sqlSelectUser = 'select * from users where id = ?';
     const sqlSelectAllCart = `select cartItems.id, book_id, title, summary, quantity, price
     from cartItems left join books on cartItems.book_id = books.id where user_id = ?`;
     const sqlSelectSelectedCart = `select cartItems.id, book_id, title, summary, quantity, price
@@ -97,14 +92,6 @@ const getCartItems = async (req, res) => {
     }
 
     try {
-        const [rowsUser] = await connection.query(sqlSelectUser, userId);
-
-        if (rowsUser.length === 0) {
-            return res.status(StatusCodes.NOT_FOUND).json({
-                message: '존재하지 않는 유저입니다.',
-            });
-        }
-
         const [rowsSelect] = await connection.query(sqlSelectCart, values);
 
         if (rowsSelect.length > 0) {
@@ -126,20 +113,22 @@ const getCartItems = async (req, res) => {
 
 const deleteCartItem = async (req, res) => {
     const connection = await conn.getConnection();
-    const { id } = req.params;
+    const bookId = req.params.id;
+    const userId = req.userId;
+    const values = [bookId, userId];
 
-    const sqlDeleteCart = 'delete from cartItems where id = ?';
+    const sqlDeleteCart = 'delete from cartItems where id = ? and user_id = ?';
 
     try {
-        const [rowsDelete] = await connection.query(sqlDeleteCart, id);
+        const [rowsDelete] = await connection.query(sqlDeleteCart, values);
 
         if (rowsDelete.affectedRows > 0) {
             return res.status(StatusCodes.OK).json({
-                message: '장바구니에서 도서가 삭제되었습니다.',
+                message: '장바구니 도서가 삭제되었습니다.',
             });
         } else {
-            return res.status(StatusCodes.NOT_FOUND).json({
-                message: '존재하지 않는 장바구니 도서입니다.',
+            return res.status(StatusCodes.BAD_REQUEST).json({
+                message: '장바구니 도서를 삭제할 수 없습니다.',
             });
         }
     } catch (err) {
